+++ views_lnurl.py
@@ -1,64 +1,25 @@
 import base64
-import hmac
 from http import HTTPStatus
-from io import BytesIO
 
-import bolt11
-from embit import compact
-from fastapi import APIRouter, HTTPException, Query, Request
-from lnbits.core.services import create_invoice, pay_invoice
+from fastapi import HTTPException, Query, Request
+
+from lnbits import bolt11
+from lnbits.core.services import create_invoice
+from lnbits.core.views.api import pay_invoice
 from lnbits.utils.exchange_rates import fiat_amount_as_satoshis
+from loguru import logger
 
+from . import lnurldevice_ext
 from .crud import (
     create_lnurldevicepayment,
     get_lnurldevice,
     get_lnurldevicepayment,
-    get_lnurldevicepayment_by_p,
     update_lnurldevicepayment,
 )
+from fastapi.responses import JSONResponse
+from .helpers import register_atm_payment, xor_decrypt
 
-lnurldevice_lnurl_router = APIRouter()
-
-
-def xor_decrypt(key, blob):
-    s = BytesIO(blob)
-    variant = s.read(1)[0]
-    if variant != 1:
-        raise RuntimeError("Not implemented")
-    # reading nonce
-    val = s.read(1)[0]
-    nonce = s.read(val)
-    if len(nonce) != val:
-        raise RuntimeError("Missing nonce bytes")
-    if val < 8:
-        raise RuntimeError("Nonce is too short")
-
-    # reading payload
-    val = s.read(1)[0]
-    payload = s.read(val)
-    if len(payload) > 32:
-        raise RuntimeError("Payload is too long for this encryption method")
-    if len(payload) != val:
-        raise RuntimeError("Missing payload bytes")
-    hmacval = s.read()
-    expected = hmac.new(
-        key, b"Data:" + blob[: -len(hmacval)], digestmod="sha256"
-    ).digest()
-    if len(hmacval) < 8:
-        raise RuntimeError("HMAC is too short")
-    if hmacval != expected[: len(hmacval)]:
-        raise RuntimeError("HMAC is invalid")
-    secret = hmac.new(key, b"Round secret:" + nonce, digestmod="sha256").digest()
-    payload = bytearray(payload)
-    for i in range(len(payload)):
-        payload[i] = payload[i] ^ secret[i]
-    s = BytesIO(payload)
-    pin = compact.read_from(s)
-    amount_in_cent = compact.read_from(s)
-    return str(pin), amount_in_cent
-
-
-@lnurldevice_lnurl_router.get(
+@lnurldevice_ext.get(
     "/api/v1/lnurl/{device_id}",
     status_code=HTTPStatus.OK,
     name="lnurldevice.lnurl_v1_params",
@@ -75,7 +36,7 @@ async def lnurl_v1_params(
     return await lnurl_params(request, device_id, p, atm, gpio, profit, amount)
 
 
-@lnurldevice_lnurl_router.get(
+@lnurldevice_ext.get(
     "/api/v2/lnurl/{device_id}",
     status_code=HTTPStatus.OK,
     name="lnurldevice.lnurl_v2_params",
@@ -126,18 +87,18 @@ async def lnurl_params(
 
         # Check they're not trying to trick the switch!
         check = False
-        if device.switches:
-            for switch in device.switches:
+        if device.extra and "atm" not in device.extra:
+            for extra in device.extra:
                 if (
-                    switch.pin == int(pin)
-                    and switch.duration == int(duration)
-                    and bool(switch.variable) == bool(variable)
-                    and bool(switch.comment) == bool(comment)
+                    extra.pin == int(pin)
+                    and extra.duration == int(duration)
+                    and bool(extra.variable) == bool(variable)
+                    and bool(extra.comment) == bool(comment)
                 ):
                     check = True
                     continue
         if not check:
-            return {"status": "ERROR", "reason": "Switch params wrong"}
+            return {"status": "ERROR", "reason": "Extra params wrong"}
 
         lnurldevicepayment = await create_lnurldevicepayment(
             deviceid=device.id,
@@ -154,16 +115,15 @@ async def lnurl_params(
                 request.url_for(
                     "lnurldevice.lnurl_callback",
                     paymentid=lnurldevicepayment.id,
-                    variable=variable,
-                )
+                ) + f"?variable={variable}",
             ),
             "minSendable": price_msat,
             "maxSendable": price_msat,
             "metadata": device.lnurlpay_metadata,
         }
-        if comment:
+        if comment == True:
             resp["commentAllowed"] = 1500
-        if variable:
+        if variable == True:
             resp["maxSendable"] = price_msat * 360
         return resp
 
@@ -180,43 +140,23 @@ async def lnurl_params(
         await fiat_amount_as_satoshis(float(amount_in_cent) / 100, device.currency)
         if device.currency != "sat"
         else amount_in_cent
-    ) * 1000
+    )
 
     if atm:
-        if device.device != "atm":
-            return {"status": "ERROR", "reason": "Not ATM device."}
-        price_msat = int(price_msat * (1 - (device.profit / 100)) / 1000)
-        new_lnurldevicepayment = await get_lnurldevicepayment_by_p(p)
-        if new_lnurldevicepayment:
-            if new_lnurldevicepayment.payload == new_lnurldevicepayment.payhash:
-                return {"status": "ERROR", "reason": "Payment already claimed"}
-        try:
-            lnurldevicepayment = await create_lnurldevicepayment(
-                deviceid=device.id,
-                payload=p,
-                sats=price_msat * 1000,
-                pin=pin,
-                payhash="payment_hash",
-            )
-        except Exception:
-            return {"status": "ERROR", "reason": "Could not create ATM payment."}
+        lnurldevicepayment = await register_atm_payment(device, p)
         if not lnurldevicepayment:
             return {"status": "ERROR", "reason": "Could not create ATM payment."}
         return {
             "tag": "withdrawRequest",
-            "callback": str(
-                request.url_for(
-                    "lnurldevice.lnurl_callback",
-                    paymentid=lnurldevicepayment.id,
-                    variable=None,
-                )
-            ),
+            "callback": str(request.url_for(
+                "lnurldevice.lnurl_callback", paymentid=lnurldevicepayment.id
+            )),
             "k1": p,
             "minWithdrawable": price_msat * 1000,
             "maxWithdrawable": price_msat * 1000,
             "defaultDescription": f"{device.title} - pin: {lnurldevicepayment.pin}",
         }
-    price_msat = int(price_msat * ((device.profit / 100) + 1) / 1000)
+    price_msat = int(price_msat * ((device.profit / 100) + 1))
 
     lnurldevicepayment = await create_lnurldevicepayment(
         deviceid=device.id,
@@ -229,28 +169,24 @@ async def lnurl_params(
         return {"status": "ERROR", "reason": "Could not create payment."}
     return {
         "tag": "payRequest",
-        "callback": str(
-            request.url_for(
-                "lnurldevice.lnurl_callback",
-                paymentid=lnurldevicepayment.id,
-                variable=None,
-            )
-        ),
+        "callback": str(request.url_for(
+            "lnurldevice.lnurl_callback", paymentid=lnurldevicepayment.id
+        )),
         "minSendable": price_msat * 1000,
         "maxSendable": price_msat * 1000,
         "metadata": device.lnurlpay_metadata,
     }
 
 
-@lnurldevice_lnurl_router.get(
-    "/api/v1/lnurl/cb/{paymentid}/{variable}",
+@lnurldevice_ext.get(
+    "/api/v1/lnurl/cb/{paymentid}",
     status_code=HTTPStatus.OK,
     name="lnurldevice.lnurl_callback",
 )
 async def lnurl_callback(
     request: Request,
     paymentid: str,
-    variable: str,
+    variable: str = Query(None),
     amount: int = Query(None),
     comment: str = Query(None),
     pr: str = Query(None),
@@ -307,10 +243,7 @@ async def lnurl_callback(
         payment_hash, payment_request = await create_invoice(
             wallet_id=device.wallet,
             amount=int(amount / 1000),
-            memo=(
-                f"{device.id} pin {lnurldevicepayment.pin} "
-                f"({lnurldevicepayment.payload} ms)"
-            ),
+            memo=f"{device.id} pin {lnurldevicepayment.pin} ({lnurldevicepayment.payload} ms)",
             unhashed_description=device.lnurlpay_metadata.encode(),
             extra={
                 "tag": "Switch",
